photo-processor/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
websockets==14.1
aiofiles==24.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
redis==5.2.1
pillow==11.0.0
numpy==1.26.4
rawpy==0.19.1

# Celery for background task processing
celery[redis]==5.4.0
flower==2.0.1
kombu==5.4.2

# Database
aiosqlite==0.21.0

# Development tools
watchdog==5.0.3

# Additional dependencies
requests==2.32.3

# Computer Vision
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
scikit-image==0.24.0
scipy==1.14.1

# PyTorch for AI components
torch==2.1.0
torchvision==0.16.0